Add verification_metrics.hpp
