+++ dawn/src/driver-includes/verification_metrics.hpp
@@ -1,0 +1,10 @@
+#pragma once
+
+struct VerificationMetrics {
+  int iteration;
+  bool isValid;
+  double maxRelErr;
+  double minRelErr;
+  double maxAbsErr;
+  double minAbsErr;
+};
